web/requirements.txt
flask
pyjwt
flask_sqlalchemy
requests
flask-socketio
eventlet
pyopenssl
python-dotenv